+++ tests/testthat/test-get_schedule.R
@@ -14,3 +14,9 @@ test_that("errors if schedule is not a MaestroSchedule", {
     get_schedule(iris)
   }, regexp = "Schedule must be an object")
 })
+
+test_that("get_schedule works with DAG schedules", {
+  schedule <- build_schedule(test_path("test_pipelines_dags_good"))
+  expect_snapshot(get_schedule(schedule))
+}) |>
+  suppressMessages()