+++ tests/testthat/test_pipelines/test_pipeline_outputs_good.R
@@ -1,0 +1,4 @@
+#' @maestroOutputs fun1 fun2 fun3
+some_outputs <- function() {
+
+}