+++ tests/testthat/test_pipelines_parse_all_bad/dags.R
@@ -3,4 +3,7 @@ with_inputs <- function() {
 
 }
 
+#' @maestroOutputs fun4 fun5
+with_outputs <- function() {
 
+}