+++ tests/testthat/test_pipelines_parse_all_bad/dags_self_reference_inputs.R
@@ -1,0 +1,6 @@
+#' @maestroInputs with_inputs
+with_inputs <- function() {
+
+}
+
+