+++ tests/testthat/test_pipelines_parse_all_good/dags.R
@@ -1,0 +1,9 @@
+#' maestroInputs fun1
+my_downstream_pipe <- function(.input) {
+
+}
+
+#' maestroOutputs my_dep_pipe
+my_upstream_pipe <- function() {
+  return(10)
+}